+++ src/app/pages/new-poem/step-2/step-2.component.ts
@@ -6,7 +6,7 @@ import { Steps } from "../steps/steps.component"
 
 // Import Local Services
 import { Forms } from "../../../services/forms.service"
-import { ReturnButton } from "../back-btn/back-btn.component"
+import { ReturnButton } from "../back-btn/button"
 
 @Component({
     selector: "app-step-2",
